+++ system/platform-core/src/main/java/org/platformlambda/core/models/EventEnvelope.java
@@ -17,6 +17,7 @@ import org.platformlambda.core.exception.AppException;
 import org.platformlambda.core.serializers.MsgPack;
 import org.platformlambda.core.serializers.SimpleMapper;
 import org.platformlambda.core.util.AppConfigReader;
+import org.platformlambda.core.util.SimpleClassScanner;
 import org.platformlambda.core.util.Utility;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -262,13 +263,27 @@ public class EventEnvelope {
      */
     public EventEnvelope restoreBodyAsPoJo() {
         if (body instanceof Map && type != null && type.contains(".") && !type.startsWith("java.")) {
-            try {
-                setBody(getBody(Class.forName(type)));
-            } catch (ClassNotFoundException e) {
-                log.warn("Unable to restore body as {} because class not found", type);
+            if (isAllowedClass(type)) {
+                try {
+                    setBody(getBody(Class.forName(type)));
+                } catch (ClassNotFoundException e) {
+                    log.warn("Unable to restore body as {} because class not found", type);
+                }
+            } else {
+                throw new IllegalArgumentException(type + " does not belong to an approved package");
             }
         }
         return this;
+    }
+
+    private boolean isAllowedClass(String className) {
+        var approvedPackages = SimpleClassScanner.getInstance().getApprovedPackages();
+        for (String p : approvedPackages) {
+            if (className.startsWith(p)) {
+                return true;
+            }
+        }
+        return false;
     }
 
     /**