+++ system/platform-core/src/main/java/org/platformlambda/core/system/AutoStart.java
@@ -21,8 +21,10 @@ package org.platformlambda.core.system;
 import org.platformlambda.core.util.AppConfigReader;
 
 import java.lang.reflect.Method;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 public class AutoStart {
+    private static final AtomicBoolean started = new AtomicBoolean(false);
 
     /**
      * This entry point decides the optimal way to start.
@@ -30,18 +32,21 @@ public class AutoStart {
      * @param args from command line, if any
      */
     public static void main(String[] args) {
-        AppConfigReader config = AppConfigReader.getInstance();
-        var springBoot = config.getProperty("spring.boot.main", "org.platformlambda.rest.RestServer");
-        try {
-            Class<?> cls = Class.forName(springBoot);
-            Method method = cls.getMethod("main", String[].class);
-            // Declare as a Spring Boot application so that AppStarter will defer loading main applications
-            AppStarter.runAsSpringBootApp();
-            // Execute BeforeApplication(s)
-            AppStarter.main(args);
-            method.invoke(null, (Object) args);
-        } catch (ReflectiveOperationException e) {
-            AppStarter.main(args);
+        if (!started.get()) {
+            started.set(true);
+            AppConfigReader config = AppConfigReader.getInstance();
+            var springBoot = config.getProperty("spring.boot.main", "org.platformlambda.rest.RestServer");
+            try {
+                Class<?> cls = Class.forName(springBoot);
+                Method method = cls.getMethod("main", String[].class);
+                // Declare as a Spring Boot application so that AppStarter will defer loading main applications
+                AppStarter.runAsSpringBootApp();
+                // Execute BeforeApplication(s)
+                AppStarter.main(args);
+                method.invoke(null, (Object) args);
+            } catch (ReflectiveOperationException e) {
+                AppStarter.main(args);
+            }
         }
     }
 }